refactor(capture): split the Sink's exclusion gate and auto-create follow-up into their own files
The auto-create pipeline pushed sink.go past the 500-LOC cap; the RC-2
exclusion gate and the ADR-0063 counterparty ensure path each carry one
concept, so each gets its own file. No behaviour change.
